Refactor and cleanup data fetching
The main reason for refactor was to decouple the actual api from
react and to treat each of the concerns (fetch client -> api auth/token
refresh handling -> request state management) as a seperate, somewhat
composable and reusable layer. This is also the reason, why are we
passing `onLogout` callback through the api(). We didn't want to turn
the api into a hook yet (at this layer) to be able to use `globalDispatch`.